[WIP] [GPU] Removed code duplication and separated concerns. 1. Created a new, generic pass for gpu host that leverages gpu dialect operations instead of directly lowering to vendor specific ones. 2. Vendor-specific lowering passes now operate on the gpu dialect-based IR operations. 3. Thread-block sizes are now passed from the generated code instead of hardcoding them in the GpuUtils.cpp
