+++ include/comet/Conversion/CMakeLists.txt
@@ -12,6 +12,7 @@ add_subdirectory(GpuToBlockedGpu)
 add_subdirectory(BlockedGpuToTriton)
 add_subdirectory(TritonToCuda)
 add_subdirectory(TritonToHIP)
+add_subdirectory(PrepareGpuHost)
 endif()
 set(LLVM_TARGET_DEFINITIONS Passes.td)
 mlir_tablegen(Passes.h.inc -gen-pass-decls -name Conversion)
