+++ lib/Conversion/CMakeLists.txt
@@ -13,4 +13,6 @@ add_subdirectory(BlockedGpuToTriton)
 add_subdirectory(GpuToTriton)
 add_subdirectory(TritonToCuda)
 add_subdirectory(TritonToHIP)
+add_subdirectory(PrepareGpuHost)
+add_subdirectory(GpuUtils)
 endif()
